experiments/fine-tune-llm/requirements.txt
openai>=1.0.0
google-genai>=1.0.0
google-cloud-storage>=2.0.0

